backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary>=2.8
pytz==2022.7.1
sqlparse==0.4.3
redgreenunittest==0.1.1
django-extensions==3.2.1
apscheduler==3.10.1